Scaffolds Harvester and MITHarvester with full harvest methods
Why these changes are being introduced:
This harvester will ultimately have a couple of Harvesters, and this begins scaffolding
them out.  This commit introduces the Harvester and MITHarvester, but with limited
functionality at this point of retrieving files locally or from S3, and invoking the
SQSClient to look for messages.

In support of this, a few utility clients for AWS services are also introduced.

Coming next will be incremental harvests for MITHarvester that relies more heavily
on SQS.

How this addresses that need:
* adds Harvester and MITHarvester classes
* adds new AWS service clients that will be reused throughout

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/GDT-84