Adds SQS abilities and MIT incremental harvest
Why these changes are being introduced:
In addition to "full" harvests that look for all zip files in a location, this harvester
also supports "incremental" harvests for MIT resources.  An incremental harvest is driven
by fetching SQS messages that contain modifications to files in S3.

How this addresses that need:
* SQSClient expanded to fetch and iterate over messages
* new ZipFileEventMessage class created to validate and work with messages
* full and incremental harvests for MIT are scaffolded to the point of processing the zip files

Side effects of this change:
None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/GDT-84