Scaffold SourceRecord normalization to MITAardvark
Why these changes are being introduced:
A primary role of this harvester is to normalize source records into MITAardvark records.  This begins to scaffold
the SourceRecord.normalize() method, with little to no actual logic for FGDC, ISO19139, GBL1, or Aardvark formats
yet, but the general pattern established.

How this addresses that need:
* SourceRecord.normalize() updated
  * gets field names from MITAardvark
  * looks for correspondign methods on SourceRecord extending class
  * produces new MITAardvark record
* more scaffolding for FGDC, ISO19139, GBL1, and Aaardvark classes which will eventually define field method classes

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/GDT-86