Skip refetching same SQS message in single harvest
Why these changes are being introduced:
The method SQSClient.get_valid_messages_iter() attempts to  exhaust the configured SQS queue and
yield these messages for the MITHarvester to work with.  However, if a record is not properly handled
during the harvest, and the harvest exceeds the SQS queue read timeout, the message is put back into
the queue and this method will pick it up again in the same harvest.

How this addresses that need:
* adds a seen list of message ids, skipping them if seen again in same harvest

Side effects of this change:
* None

Relevant ticket(s):
* None