First pass at required FGDC field methods
Why these changes are being introduced:
After establishing shared field methods on SourceRecord, this is the first pass at
defining field methods for MITAardvark required fields for the FGDC class.

Still a WIP, this begins to establish the pattern that this class will complete
and others will follow.

How this addresses that need:
* Defines required field methods on FGDC
* Demonstrates a single optional field method _dct_identifier_sm

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/GDT-86