Set gbl_suppressed_b based from event
Why these changes are being introduced:
The harvester needs a way to indicate to downstream consumers if the record it is processing is
considered deleted.  We have decided to use the Aardvark field "gbl_supressed_b" to represent this.

How this addresses that need:
* Shared field method created for gbl_suppressed_b, based on SourceRecord.event

Side effects of this change:
* Records will now have gbl_suppressed_b present

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/GDT-86