Fix unit test timezone handling
Why these changes are being introduced:
The base Harvester class relies on a utility method convert_to_utc() to convert whatever datetime object
is parsed from the CLI arguments to UTC time.  In a deployed context (or Github actions) the "local" timezone
is UTC, so this should have no effect.  From a developer machine, this is likely EST or another timezone.

Previously, the unit test had a 5 hour bump for the datetime assertion, which reflected the parsed UTC time.  When
in reality, the unit test should assert that a parsed YYYY-MM-DD date provided via the CLI matches a UTC time of that
same format.

How this addresses that need:
* Unit test does not include hours or minutes but instead asserts same year, month, date equality and UTC timestamp

Side effects of this change:
* None

Relevant ticket(s):
* None