Add TIMDEX identifier to MITAardvark file produced
Why these changes are being introduced:
A very small subset of MIT GIS zip files do not have any meaningful identifiers in the FGDC or ISO metadata.
This would add the minted identifier for TIMDEX, based heavily on the input filename, to the MITAardvark
dct_identifiers_sm array.  In many cases, this will be present already from being in the metadata, but if not
present it will add.  In either situation, it is deduped so not present twice.

How this addresses that need:
* Adds record.identifier to in get_dct_identifiers_sm() method

Side effects of this change:
* All records will have at least one value for dct_identifiers_sm

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/GDT-149