Allow skipping of EventBridge events
Why these changes are being introduced:
Mostly for debugging purposes, skipping the sending of EventBridge events for MIT harvests allows
for debugging the harvester flow and performance.  With the current ability to exclude writing external files,
sending an EventBridge event was still a non-configurable external action the harvester would take.

How this addresses that need:
* Adds CLI and harvester flag to skip sending EventBridge events

Side effects of this change:
* None

Relevant ticket(s):
* None