Build OGM repositories as pytest fixture
Why these changes are being introduced:
Incremental OGM harvests rely on the git history of the locally cloned OGM repository.
Originally, the thought was to have static git projects in the tests/fixtures directory, but
this proved problematic with comitting and including with the parent, main project.  In the
end, building these simulated OGM repositories via pytest fixture was the better option.  This
will also allow for modifying the simulated, cloned repositories more easily if edge cases
are determined.

How this addresses that need:
* Builds three, simulated OGM repositories via a pytest fixture

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/GDT-85