Use pygit2 instead of GitPython and OS git binary
Why these changes are being introduced:

OGM harvests require a fair amount of git functionality to interact with the OGM repositories.  Previously was
using GitPython for some operations, and then using the OS level git binary + subprocess for some detailed
file diff information as GitPython did not provide this.

It was discovered that pygit2 would support all operations needed, and is ultimately more low level
and flexible if needed.

How this addresses that need:
* Use pygit2 for all git operations

Side effects of this change:
* Calling context must have libgit2 available

Relevant ticket(s):
* None