Check remote repository for commits before clone
Why these changes are being introduced:
When performing an OGM incremental harvest, it is not ideal to clone an entire repository, only to
interrogate the git history locally and determine that no commits were made on or after the target date.

Though use of the Github API would be prohibitvely slow for the more complex parts of an incremental harvests,
a quick call to look for ANY commits on or after the target date is a great way to avoid cloning repositories
when it can be known in advance they will not produce any records for the harvest.

This change would allow true daily OGM harvests, as most harvests will quickly determine there were no
commits and therefore not need to clone the repository.

How this addresses that need:
* Use Github API to check for commits on or after an incremental harvest target date

Side effects of this change:
* Dramatic reduction in cloned repositories for incremental harvests.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/GDT-85