Create validator for evaluating geoshape WKT values
Why these changes are being introduced:
* This is part of an effort to add some data validation to the TIMDEX
pipeline, with an aim to catch errors with records in earlier stages
(e.g., during the harvest). This also proposes a workflow wherein
validation methods can easily be applied as decorators to field methods
for a given SourceRecord.

How this addresses that need:
* Add 'shapely' as a dependency
* Create custom warning for invalid field values
* Create decorator method for validating geoshape WKT values
   * Returns None if geoshape WKT value is invalid (cannot be parsed)

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/GDT-159