Set single URL for part of dct_references_s
Why these changes are being introduced:
It was discovered that the Aardvark format expects only a single URL for the dct_references_s
key 'http://schema.org/url', where previously we had set this as a list of URLs like
'http://schemal.org/downloadUrl'.

This also updates the hostname for where this URL points back to 'geodata.libraries.mit.edu' which
is the anticipated TIMDEX UI for geo resources.

How this addresses that need:
* Sets single URL for 'http://schema.org/url'
* Updates domain name

Side effects of this change:
* Requires an associated PR in Transmogirifer to expect a single value and not an array; forthcoming

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/GDT-68