Expose OGM configuration to normalization logic
Why these changes are being introduced:

When normalizing OGM records, it was determined that we might need some repo-by-repo specific
configurations for OGM repositories.  The established place to do this is in the OGM config YAML,
and so it makes sense to expose this config dictionary to the record classes that require it.

One of the repo specific updates that will be utilized in a future commit for GBL1 and Aardvark
classes are alternate strategies for extracting a meaningful external URL for repositories.  This
has been, optionally, added to some OGM repo configurations.

How this addresses that need:
* Adds ogm_config dictionary to SourceRecord class
* Some OGM configs get an alteranate url extraction strategy

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/GDT-109