Remove bbox requirement from MITAardvark and add WKT validator (#138)
* Remove bounding box requirement from MITAardvark

Why these changes are being introduced:
* This enables Geo-Harvester to process geospatial metadata with missing
or potentially invalid geospatial bounding box information. This change is
in preparation of the ingestion of geospatial metadata from external sources
that will likely vary in data quality. Having these fields as required may be
too restrictive.

How this addresses that need:
* Remove bounding box fields as 'required' from MITAardvark JSON schema
* Remove bounding box fields from test fixtures representing 'required' data
* Move bounding box attributes to 'optional' group in harvester.records.MITAardvark

Side effects of this change:
* Required follow-up work: Geo-Harvester will validate bounding box information when present.

---

* Create validator for evaluating geoshape WKT values

Why these changes are being introduced:
* This is part of an effort to add some data validation to the TIMDEX
pipeline, with an aim to catch errors with records in earlier stages
(e.g., during the harvest). This also proposes a workflow wherein
validation methods can easily be applied as decorators to field methods
for a given SourceRecord.

How this addresses that need:
* Add 'shapely' as a dependency
* Create custom warning for invalid field values
* Create decorator method for validating geoshape WKT values
   * Returns None if WKT value is invalid (cannot be parsed)
* Update SourceRecord classes (FGDC, ISO19139)
* Move MITAardvark format validation method into validator module
* Address comments

Side effects of this change:
* None

---

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/GDT-159