OGM normalize only include download link if single value
Why these changes are being introduced:

When normalizing OGM records, a very small subset of OGM records include an array of
values in the dct_references_s field for the URI 'http://schema.org/downloadUrl'.  When
this value is an array, and not a single value string, we cannot know which URL to select
as the only download URL we will expose in TIMDEX.  Therefore, despite the information loss,
it is better to skip a download URL for these records.

How this addresses that need:
* Checks that found download value is a string URL value, not an array

Side effects of this change:
* A small subset of records that might have had download URLs for OGM harvests will not have them

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/GDT-109