dct_format_s utilize gbl_resourceType_sm values as needed
Why these changes are being introduced:

MIT FGDC source records have an explicit file type format about half the time in the XML, which is mapped
to the Aardvark field dct_format_s as a controlled term.  However, when that value is missing, or does not
map, we are left with MITAardvark records without dct_format_s.  It was learned by talking with GIS
stakeholders that we can also rely on data about the record, coming from gbl_resourceType_sm, to inform
what the file type.  In this way, we can with high probability determine a file type for all records, even
if not explicitly stated in the XML.

How this addresses that need:
* When method get_controlled_dct_format_s_term() fails to find a controlled term, fallback on checking
the output of get_controlled_gbl_resourceType_sm_terms() for hints about the file type

Side effects of this change:
* Increased dct_format_s values for MIT FGDC records

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/GDT-195