Fetch and filter records for Alma MARC harvest
Why these changes are being introduced:

This update introduces the ability for GeoHarvester to harvest Alma MARC records.  This first update
only includes the ability to fetch and filter geospatial MARC records from S3.  The files it is targeting
are MARC XML files deposited by the TIMDEX Alma exports.

By introducing a new MITAlmaHarvester, this paves the way for a new gisalma TIMDEX source that would
sit alongside of gismit and gisogm.

It has been widely observed this is not the most efficient way to harvest a subset of Alma MARC
records that meet our geospatial requirements, but it is a path with minimal (if any) infrastructure changes
at the moment.  If and when other mechanisms to produce a smaller subset of Alma MARC records emerge,
this harvester could be updated to point at a different location, but the mechanics should remain
largely the same, potentially with some filtering still happening at this application level.

How this addresses that need:
* adds new MITAlmaHarvester harvester
* MITAlmaHarvester harvester reads through MARC files in S3 and filters to geospatial records
as defined by the harvester

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/GDT-226