Refactor classes for record sources and formats
Why these changes are being introduced:

With the introduction of Alma as a source of records, and MARC as a format, it was exposing
some pain points in the previous architecture of classes that define a SourceRecord.

The logic needed to derive a normalized, MIT Aardvark file is a combination of where it came from
(source) and the metadata format of the original record.  Previously, there was a bit of leakage
between source and format, where format classes like FGDC may have some logic specific to only
OGM harvests.

By extending two classes, one specific to the source and one for the format, the logic can be correctly
encapsulated.  This also removes some conditional if/else and match logic in the SourceRecord class that
was specific to the record source.

How this addresses that need:
* 'formats' directory created under 'records' populated with metadata format classes, e.g. FGDC, ISO19139, etc.
* 'sources' directory created under 'records' populated with classes that define logic specific to the source, further
extending format classes
* each source, currenrly MIT, OGM, and Alma, all get composite SourceRecord classes of source + format that are specific to their harvest
* composite source + format classes have field methods defined like schema_provider_s or dct_references_s

Side effects of this change:
* None - inputs/outpus are the same, just code refactor

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/GDT-227