Required field methods for MARC transform
Why these changes are being introduced:
This begins the process of filling out the MARC transform to MITAardvark for
the required fields only.  This work also establishes some patterns and helpers
for the MARC and MarcalyxSourceRecord classes that will benefit optional
fields to come.

How this addresses that need:
* Creates helpers on MarcalyxSourceRecord to work with parsed marcalyx records
* Defines required fields for MARC to MITAardvark normalization

Side effects of this change:
* AlmaMARC and MARC are capable of producing valid MITAardvark records sans optional
fields

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/GDT-227