Use tag 336 and 007/00 for gbl_resourceClass_sm
Why these changes are being introduced:

It was observed that the previous mapping of 336$a had a typo in the mapping, and this typo
revealed that relying on tag 336$a for the ResourceClass value alone would not properly
identify true Maps.  It does a good job of telling us Dataset vs Imagery (visual resource),
but we can then use tag 007/00 to get a sense of Map vs Imagery.

How this addresses that need:
* Utilize tag 007/00 in _gbl_resourceClass_sm() field method

Side effects of this change:
* More accurate high level ResourceClass values

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/GDT-227