Gracefully exit harvest when no records
Why these changes are being introduced:

Before this change, if no records were discovered for harvesting, the full
pipeline of steps were still executed, just with an empty iterator.  For the most
part this was inconsequential, except that an empty output file was still created
for the normalized MIT Aardvark records.  If the parameters of a harvest do not
discover any records for harvesting, this output file should not be created.

How this addresses that need:
* In the primary, shared harvest() method, a check is added after get_source_records()
that confirms the records iterator has records.  If not, the harvest is exited early.
* Harvest results moved to property so it can be used in multiple places.

Side effects of this change:
* When harvests do not have records, any empty output file will not be created.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/GDT-267