Use Github RSS vs API for incremental OGM harvest
Why these changes are being introduced:

It was learned that, without a Github API token, unauthenticated requests to
the Github API were failing immediately when invoked from deployed ECS containers,
getting a 403 rate limiting error.  From this it was learned that Github also provides
an Atom RSS feed for recent commits to a repository.  This RSS feed contains the date
of the commit, and does not suffer from the same rate limiting restrictions, and is
therefore a good replacement for using the API to get this information.

How this addresses that need:
* Remove Github API token and API calls everywhere
* Update OGMHarvester to query Atom RSS endpoint for recent commits instead of API

Side effects of this change:
* Incremental/daily harvests run successfully in deployed ECS context

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/GDT-269